Former-commit-id: 309062e4095b4950608a607decf95f7cff2f4dc1

+++ ifs-data-service/src/main/java/org/innovateuk/ifs/assessment/transactional/AssessmentServiceImpl.java
@@ -113,7 +113,7 @@ public class AssessmentServiceImpl extends BaseTransactionalService implements A
     }
 
     @Override
-    public ServiceResult<Long> countByStateAndCompetition(AssessmentStates state, long competitionId) {
+    public ServiceResult<Integer> countByStateAndCompetition(AssessmentStates state, long competitionId) {
         return serviceSuccess(assessmentRepository.countByActivityStateStateAndTargetCompetitionId(state.getBackingState(), competitionId));
     }
